refactor: removida a redundância de pacotes com arquivos únicos
