+++ HotelAPI/src/main/java/br/com/totvs/hotel/model/PessoaModel.java
@@ -1,0 +1,43 @@
+package br.com.totvs.hotel.model;
+
+import br.com.totvs.hotel.model.EnderecoModel;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDate;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Entity
+@Table(name = "table_pessoa")
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
+public abstract class PessoaModel {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(nullable = false)
+    private String nome;
+
+    @Column(nullable = false)
+    private String sobrenome;
+
+    @Column(nullable = false)
+    private String rg;
+
+    @Column(nullable = false)
+    private String cpf;
+
+    @Column(nullable = false)
+    private LocalDate dataNascimento;
+
+    @OneToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "id_endereco")
+    private EnderecoModel endereco;
+
+}
