+++ HotelAPI/src/main/java/br/com/totvs/hotel/model/QuartoModel.java
@@ -1,0 +1,52 @@
+package br.com.totvs.hotel.model;
+
+import br.com.totvs.hotel.enumeration.quarto.CategoriaQuarto;
+import br.com.totvs.hotel.enumeration.quarto.SituacaoQuarto;
+import br.com.totvs.hotel.model.EstadiaModel;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Entity
+@Table(name = "table_quarto")
+public class QuartoModel {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(nullable = false)
+    private Integer numero;
+
+    @Column(nullable = false)
+    private String descricao;
+
+    @Column(nullable = false)
+    private Double precoHora;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private CategoriaQuarto categoria;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private SituacaoQuarto situacao;
+
+    @ElementCollection
+    @CollectionTable(name = "table_quarto_imagens", joinColumns = @JoinColumn(name = "id_quarto"))
+    @Column(name = "url", length = 65535)
+    private List<String> imagens;
+
+    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
+    @JoinColumn(name = "id_quarto")
+    private List<EstadiaModel> estadias = new ArrayList<EstadiaModel>();
+
+}
