+++ HotelAPI/src/main/java/br/com/totvs/hotel/repository/EnderecoRepository.java
@@ -1,0 +1,10 @@
+package br.com.totvs.hotel.repository;
+
+import br.com.totvs.hotel.model.EnderecoModel;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EnderecoRepository extends JpaRepository<EnderecoModel, Long> {
+
+}
