--- HotelAPI/src/main/java/br/com/totvs/hotel/validation/Age.java
@@ -1,21 +1,0 @@
-package br.com.totvs.hotel.validation;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = AgeValidator.class)
-public @interface Age {
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-    String message() default "idade informada inválida";
-    int minimum() default 0;
-    int maximum() default 100;
-
-}
